metric sync_delta and sync_all with datacenter label
