+++ server/server/session/src/main/java/com/alipay/sofa/registry/server/session/scheduler/timertask/CacheCountTask.java
@@ -180,4 +180,10 @@ public class CacheCountTask {
       }
     }
   }
+
+  public static void main(String[] args) {
+    List<String> notMulti = Lists.newArrayListWithExpectedSize(1000);
+    notMulti.add("aaa");
+    System.out.println(notMulti.size());
+  }
 }
